Refactor acoustid json parsing code
